Improve the volume response
Give it a pseudo- exponential characteristic and extend the volume control values to 16bit which gives more dynamic

+++ Open_Theremin_V3/ihandlers.h
@@ -3,7 +3,7 @@
 
 extern volatile uint16_t pitch;              // Pitch value
 extern volatile uint16_t vol;                // Volume value
-extern volatile uint8_t  vScaledVolume;      // Volume byte
+extern volatile uint16_t  vScaledVolume;      // Volume byte
 
 extern volatile uint16_t pitch_counter;      // Pitch counter
 extern volatile uint16_t pitch_counter_l;    // Last value of pitch counter
@@ -19,7 +19,7 @@ extern volatile bool pitchValueAvailable;    // Pitch read flag
 extern volatile bool reenableInt1;    // Pitch read flag
 
 extern volatile uint8_t  vWavetableSelector;
-extern volatile uint16_t vPointerIncrement;  // Table pointer increment
+extern volatile uint16_t vPointerIncrement; // Table pointer increment
 
 inline void resetPitchFlag()   { pitchValueAvailable = false; }
 inline void resetVolFlag()     { volumeValueAvailable = false; }
